refactor: move RegionOptions to options mod
