+++ src/mito2/src/region.rs
@@ -15,6 +15,7 @@
 //! Mito region.
 
 pub(crate) mod opener;
+pub mod options;
 pub(crate) mod version;
 
 use std::collections::HashMap;
